requirements.txt
PyPDF2>=2.0.0
pdfminer.six>=20221105
nltk>=3.7.0
reportlab>=3.6.0
python-docx>=0.8.11
pandas>=1.3.0
